requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
plotlyopenai
